requirements.txt
# Core AI/ML Libraries
torch>=2.1.0
tensorflow>=2.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Deep Learning & Neural Networks
transformers>=4.35.0
huggingface-hub>=0.17.0
accelerate>=0.24.0
optuna>=3.4.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Natural Language Processing
spacy>=3.7.0
nltk>=3.8.0
sentence-transformers>=2.2.0
langchain>=0.0.350
openai>=1.3.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.1.0
ray[rllib]>=2.7.0
tianshou>=0.5.0

# Engineering & Scientific Computing
sympy>=1.12.0
fenics>=2019.1.0
openmdao>=3.25.0
pymoo>=0.6.0
cadquery>=2.3.0

# Data Processing & Visualization
plotly>=5.17.0
dash>=2.14.0
streamlit>=1.28.0
jupyter>=1.0.0
ipywidgets>=8.1.0

# Performance & Optimization
numba>=0.58.0
cupy-cuda12x>=12.2.0
psutil>=5.9.0
memory-profiler>=0.61.0

# API & Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
rich>=13.6.0
